requirements.txt
dateparser
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
numpy>=1.19.3
openpyxl
pandas
pydantic
python-dotenv
xlrd
